Se creeo controlador y validaciones para gestión de empleados; implementar rutas API para crear, buscar, actualizar y eliminar empleados.

+++ database/migrations/2026_03_29_034530_create_nominas_mensuales_table.php
@@ -24,8 +24,8 @@ return new class extends Migration
                 ->cascadeOnDelete();
             // Empleado al que pertenece la nómina
 
-            $table->integer('mes'); // Mes de la nómina
-            $table->integer('anio'); // Año de la nómina
+            $table->tinyInteger('mes');
+            $table->integer('anio');
 
             $table->decimal('horas_total', 5, 2);
             // Total de horas trabajadas en el mes
